Zmena teploty na FLOAT

+++ Meteostanice/Meteostanice/src/sendViaWifi.h
@@ -1,3 +1,3 @@
 void setupWifiCon();
-void sendDataViaWifi(int teplota, int vlhkost, int tlak, int co2, bool zda_prsi);
+void sendDataViaWifi(float teplota, float vlhkost, int tlak, int co2, bool zda_prsi);
 void sendAlarmViaWifi();
